fixed a8b338d from https://gitee.com/zgit2021/communication_dsoftbus/pulls/1278 fix code warning

+++ core/connection/tcp/src/softbus_tcp_connect_manager.c
@@ -217,7 +217,7 @@ static char *RecvData(const ConnPktHead *head, int32_t fd, int32_t len)
                 "receiveData: error occurred![recvLen=%d][len=%d]", recvLen, len);
             goto EXIT;
         }
-        recvLen += n;
+        recvLen += (uint32_t)n;
     }
     return data;
 EXIT:
